+++ domains/Discussions/routes.php
@@ -1,6 +1,7 @@
 <?php
 
 use Domains\Discussions\Controllers\AnswersStoreController;
+use Domains\Discussions\Controllers\QuestionsDeleteController;
 use Domains\Discussions\Controllers\QuestionsStoreController;
 use Domains\Discussions\Controllers\QuestionsUpdateController;
 use Domains\Discussions\Controllers\QuestionsViewController;
@@ -15,6 +16,11 @@ Route::group(['middleware' => 'auth'], function () {
     Route::patch('/questions/{questionId}', [
         'as' => 'questions.update',
         'uses' => QuestionsUpdateController::class,
+    ]);
+
+    Route::delete('/questions/{questionId}', [
+        'as' => 'questions.delete',
+        'uses' => QuestionsDeleteController::class
     ]);
 
     Route::post('/questions/{questionId}/answers', [