+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to `laravel-portugal/api` will be documented in this file
 
 ### Added
 
+- An authenticated user can delete a question (#30)
 - A guest or an authenticated user can see details of a question (#48)
 
 ### Changed